Move profiles tests to a different module
We start to have a lot of verbose profiles tests, let's move them to a
different test file.
